parse graphql schema on api save
